ml-service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.6
pydantic-settings==2.7.1
python-multipart==0.0.20
pdfplumber==0.11.5
python-docx==1.1.2
sentence-transformers==3.3.1
numpy==2.2.1
